COM-2048 replace View by ScrollView & flatList by events.map, and modify helpers/dates
