+++ src/core/helpers/utils.ts
@@ -1,20 +1,5 @@
-import { DAYS, MONTHS } from '../data/constants';
-
 export const formatPhone = (phoneNumber : any) => (phoneNumber
   ? phoneNumber.replace(/^(\d{2})(\d{2})(\d{2})(\d{2})(\d{2})$/, '$1 $2 $3 $4 $5') : '');
-
-export const formatDate = (date: Date) => {
-  const day = date.getDay();
-  const month = date.getMonth();
-  return `${DAYS[day]} ${date.getDate()} ${MONTHS[month]}`;
-};
-
-export const formatTime = (date: Date) => `${date.getHours()}:${displayMinutes(date)}`;
-
-export const displayMinutes = (date: Date) => {
-  if (date.getMinutes() < 10) return '0'.concat(date.getMinutes().toString());
-  return date.getMinutes();
-};
 
 export const capitalizeFirstLetter = (s: string) => `${s.charAt(0).toUpperCase()}${s.substr(1)}`;
