+++ src/screens/TimeStampingProfile/index.tsx
@@ -1,13 +1,21 @@
 import React, { useContext, useEffect, useState, useCallback } from 'react';
-import { Text, View, FlatList } from 'react-native';
+import { Text, View, ScrollView } from 'react-native';
 import commonStyle from '../../styles/common';
 import Events from '../../api/Events';
 import { Context as AuthContext } from '../../context/AuthContext';
 import { INTERVENTION } from '../../core/data/constants';
-import { formatTime, formatDate, capitalizeFirstLetter, formatWordToPlural } from '../../core/helpers/utils';
+import { capitalizeFirstLetter, formatWordToPlural } from '../../core/helpers/utils';
+import { formatTime, formatDate } from '../../core/helpers/dates';
 import TimeStampingCell from '../../components/TimeStampingCell';
 import styles from './styles';
 import { EventType } from '../../types/EventType';
+
+const renderEvent = (event: EventType) => (
+  <View key={event._id}>
+    <TimeStampingCell event={event} />
+    <View style={styles.separator} />
+  </View>
+);
 
 const TimeStampingProfile = () => {
   const [currentDate, setCurrentDate] = useState<Date>(new Date());
@@ -15,9 +23,10 @@ const TimeStampingProfile = () => {
 
   const { loggedUser } = useContext(AuthContext);
 
-  setInterval(() => {
-    setCurrentDate(new Date());
-  }, 60000);
+  useEffect(() => {
+    const interval = setInterval(() => { setCurrentDate(new Date()); }, 60000);
+    return () => clearInterval(interval);
+  }, []);
 
   const fetchInterventions = useCallback(async () => {
     try {
@@ -40,12 +49,8 @@ const TimeStampingProfile = () => {
     if (loggedUser?._id) fetchInterventions();
   }, [loggedUser, fetchInterventions]);
 
-  const renderEvent = (event: EventType) => <TimeStampingCell event={event} />;
-
-  const renderSeparator = () => <View style={styles.separator} />;
-
   return (
-    <View style={styles.screen}>
+    <ScrollView style={styles.screen}>
       <Text style={commonStyle.title}>Horodatage</Text>
       <View style={styles.container}>
         <View>
@@ -56,9 +61,8 @@ const TimeStampingProfile = () => {
           <Text style={styles.textIntervention}>{events.length} {formatWordToPlural(events, 'intervention')}</Text>
         </View>
       </View>
-      <FlatList data={events} keyExtractor={event => event._id} renderItem={({ item }) => renderEvent(item)}
-        ItemSeparatorComponent={renderSeparator} />
-    </View>
+      {events.map(event => renderEvent(event))}
+    </ScrollView>
   );
 };
